refactor(API): create middlewares
